bento-cluster: CLUSTER-5. bento-cluster should not start if any ports are taken.
This commit modifies bento-cluster so that it does not start if any of the
configured ports are taken. To facilitate this, a new tool was added which
checks that all the ports are open. If the check fails, the bento script refuses
to start the clusters.

Signed-off-by: Aaron Kimball <aaron@wibidata.com>
commit: https://github.com/kijiproject/bento-cluster/commit/1cdd7363079eaffe313d97bdee93be3a7c5c1142